refactor: break out concert card from ConcertList to its own reusable component
